Self-hosted extensions (#3)
**User-Facing Changes**
- Custom extensions can be bundled into a self-hosted Docker deployment

**Description**
In a similar vein to overriding the default layout in a self-hosted
Docker deployment, this adds support for bind-mounting a directory
containing `.foxe` files and a `manifest.json` (described in README.md)
when the Docker container is started to statically embed the manifest
into the served Foxglove Studio. The manifest.json is read at container
start and injected into index.html, while the foxe extension files are
served over HTTP(s) from `/extensions/`.