CBUF service calls (#5)
**User-Facing Changes**
- Call Service panel now works with servers that advertise `"cbuf"`
encoded services

**Description**
There are a few places in the WebSocket protocol where encodings or type
names are not explicitly specified and need to be inferred using
assumptions from the ROS world. One is that the schema encoding can be
assumed given the message encoding (json -> jsonschema, cdr -> ros2idl,
etc) and the other is the service request and response type names are
the service `"type"` suffixed with `_Request` and `_Response`.

A more thorough fix would involve breaking changes to ws-protocol to
explicitly specifying schema encodings and request and response type
names. This change continues the status quo but with special handling
for cbuf, similar to how ROS is special-cased.