Kiosk mode (#6)
**User-Facing Changes**
* Added kiosk mode to disable most UI and interactivity with
`?kiosk=true`

**Description**

This adds support for a new query parameter, `?kiosk=true`, that removes
much of the UI including topbar buttons, sidebars, panel resizing and
drag&drop, layout editing keyboard shortcuts, and 3D panel overlay
tools. The intention is to provide a locked down view for embedding in a
webpage, building an "operator dashboard", or other use cases a
streamlined user interface is desired or where the layout should not be
modified.